+++ README.MD
@@ -8,3 +8,11 @@ Version: 0.0.1
 Страница, на которой можно лайкать, удалять, приближать карточки.
 
 Также добавлять новые карточки, редактировать данные профиля (при редактировании встроена валидация).
+
+
+## Команды
+npm run build - команда сборки проекта
+
+npm run dev - команда запуска сборки, открывает сайт в браузере
+
+npm run deploy — команда для размещения проекта на сервере
